Move some code into video tracker shared.
