+++ plugins/videotrackershared/CMakeLists.txt
@@ -24,4 +24,5 @@ set(OSVR_VIDEOTRACKERSHARED_SOURCES_IO
 set(OSVR_VIDEOTRACKERSHARED_SOURCES_HDKDATA
     "${CMAKE_CURRENT_SOURCE_DIR}/HDKData.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/HDKData.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/LoadCalibration.h"
     CACHE INTERNAL "" FORCE)
